+++ script/Deploy.s.sol
@@ -53,7 +53,7 @@ contract DeployScript is Script, BaseInputs {
             address(treasury),
             address(state)
         );
-        p2pSwap = new P2PSwap(address(evvm), address(staking), admin);
+        p2pSwap = new P2PSwap(address(evvm), address(staking), address(state), admin);
 
         vm.stopBroadcast();