+++ test/fuzz/p2pSwap/fuzzTest_P2PSwap_makeOrder.t.sol
@@ -116,6 +116,7 @@ contract fuzzTest_P2PSwap_makeOrder is Test, Constants {
             COMMON_USER_NO_STAKER_1.PrivateKey,
             Erc191TestBuilder.buildMessageSignedForMakeOrder(
                 evvm.getEvvmID(),
+                address(p2pSwap),
                 input.nonceP2PSwap,
                 tokenA,
                 tokenB,