+++ CHANGELOG.md
@@ -31,7 +31,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **NameService**:
   - Update variable name `clowNumber` to `lockNumber` and `expireDate` to `expirationDate` for better clarity
   - Updated `NameServiceStructs` to be a library instead of an abstract contract for better modularity and reuse across services
-  - Implemented `State.sol` for nonce validation and signature verification in `makeOffer` and `withdrawOffer`, replacing previous service-specific nonce management and signature utilities
+  - Implemented `State.sol` for nonce validation and signature verification replacing previous service-specific nonce management and signature utilities
+- **P2PSwap**:
+  - Implemented `State.sol` for nonce validation and signature verification replacing previous service-specific nonce management and signature utilities
 
 ### Removed